updated javascript to be module based
