+++ _includes/layouts/base.njk
@@ -104,6 +104,7 @@
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/js/bootstrap.bundle.min.js" integrity="sha256-wMCQIK229gKxbUg3QWa544ypI4OoFlC2qQl8Q8xD8x8=" crossorigin="anonymous"></script>
     <script async src="https://js.sitesearch360.com/plugin/bundle/2941.js"></script>
+    <script src="{{ "/js/app.js" | url }}" type="module"></script>
     <!-- Global site tag (gtag.js) - Google Analytics -->
     <script async src="https://www.googletagmanager.com/gtag/js?id=G-23YFD2K57P"></script>
     <script>
