+++ js/app.js
@@ -1,0 +1,7 @@
+import jkBlog from "./module.blog.js";
+import jkTagList from "./module.tag-list.js";
+
+window.jk = {
+    blog: jkBlog,
+    tagList: jkTagList
+};
